requirements.txt
Flask==3.1.0
line-works-sdk==3.5
